move types to Api directory
